refactor: Update authentication controller to set access token as HTTP-only cookie
This commit updates the authentication controller to set the access token as an HTTP-only cookie. The access token is generated using the user's information and is stored in a cookie with the name "access_token". The cookie has a maximum age of 15 minutes and is secure, meaning it can only be accessed over HTTPS. This change improves security by preventing client-side code from accessing the access token directly.

Note: This message is based on the provided code changes and recent repository commits.